Changes for #Issue56
