+++ src/main/java/org/hobbit/utils/rdf/HashingRDFProperties.java
@@ -1,0 +1,36 @@
+package org.hobbit.utils.rdf;
+
+import java.util.stream.Collectors;
+
+import org.apache.commons.codec.digest.DigestUtils;
+import org.apache.jena.ext.com.google.common.collect.Streams;
+import org.apache.jena.rdf.model.Statement;
+import org.apache.jena.rdf.model.StmtIterator;
+
+public class HashingRDFProperties {
+	/**
+     * Computes a SHA1 hash of sorted and serialized resource properties.
+     *
+     * @param resource
+     *            the resource object for which this method computes the hash.
+     * @return the computed hash.
+     */
+    public static String hashProperties(StmtIterator statements) {
+        String s = Streams.stream(statements)
+                .map(HashingRDFProperties::serializeStatement)
+                .sorted()
+                .collect(Collectors.joining());
+
+        return DigestUtils.sha1Hex(s);
+    }
+
+    private static String serializeStatement(Statement stmt) {
+        StringBuilder s = new StringBuilder();
+        s.append(stmt.getPredicate().visitWith(HashingRDFVisitor.instance));
+        s.append(" ");
+        s.append(stmt.getObject().visitWith(HashingRDFVisitor.instance));
+        s.append("\n");
+        return s.toString();
+    }
+
+}
